feat:심화 과제 hook 분리
